Fixed an undefined error when joining a network
I found a case where an undefined error would be given if Apple told us connecting was done, but did not supply an error and the connected SSID was different from what we were trying to connect to. At least in iOS 14, the connected SSID can be nil in this case, and it does in fact mean it could not connect, which is also what it says in the OS provided alert to the user.